+++ src/main/java/com/crudapplication2/application2/service/TaskService.java
@@ -5,7 +5,7 @@ import com.crudapplication2.application2.model.response.msResponse.MSResponse;
 
 public interface TaskService {
     MSResponse addTask(TaskRequestdto task);
-    MSResponse updateTask(long id, TaskRequestdto task);
+    MSResponse updateTask(int id, TaskRequestdto task);
     MSResponse getAllTasks();
-    MSResponse deleteTask(long id);
+    MSResponse deleteTask(int id);
 }
